python-scenarios/functions/requirements.txt
pulsar-client 
avro-python3
fastavro
openai